add inventory service to use open feign

+++ src/main/java/guru/sfg/beer/inventory/service/services/AllocationServiceImpl.java
@@ -72,12 +72,16 @@ public class AllocationServiceImpl implements AllocationService {
                 beerOrderLine.setQuantityAllocated(orderQty);
                 beerInventory.setQuantityOnHand(inventory);
 
+                beerInventoryRepository.save(beerInventory);
+
             } else if(inventory > 0 ) { // partial allocation
                 beerOrderLine.setQuantityAllocated(allocatedQty + inventory);
                 beerInventory.setQuantityOnHand(0);
             }
 
-            beerInventoryRepository.delete(beerInventory);
+            if(beerInventory.getQuantityOnHand() == 0 ) {
+                beerInventoryRepository.delete(beerInventory);
+            }
 
         });
     }
